Changed height and width parameters to uint16_t (from uint8_t)
Which might explain why I was having problems printing images > 256px wide/high!

+++ src/ThermalPrinter.h
@@ -48,7 +48,7 @@ namespace PlayfulTechnology {
 			void setBarcodeHeight(int val);
 
 			// Images
-			void printBitmap(uint8_t w, uint8_t h,  const uint8_t *bitmap);
+			void printBitmap(uint16_t w, uint16_t h,  const uint8_t *bitmap);
 
 		private:
 			Stream *serial;
